pr-sug: update with sep classes
