+++ src/Arcus.Observability.Tests.Integration/Serilog/ApplicationInsightsSinkEventTests.cs
@@ -1,0 +1,137 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Arcus.Observability.Correlation;
+using Arcus.Observability.Tests.Core;
+using Microsoft.Azure.ApplicationInsights;
+using Microsoft.Azure.ApplicationInsights.Models;
+using Microsoft.Extensions.Logging;
+using Serilog;
+using Xunit;
+using Xunit.Abstractions;
+using ILogger = Microsoft.Extensions.Logging.ILogger;
+
+namespace Arcus.Observability.Tests.Integration.Serilog 
+{
+    public class ApplicationInsightsSinkEventTests : ApplicationInsightsSinkTests
+    {
+        public ApplicationInsightsSinkEventTests(ITestOutputHelper outputWriter) : base(outputWriter)
+        {
+        }
+
+        [Fact]
+        public async Task LogEvent_SinksToApplicationInsights_ResultsInEventTelemetry()
+        {
+            // Arrange
+            string eventName = BogusGenerator.Finance.AccountName();
+            using (ILoggerFactory loggerFactory = CreateLoggerFactory())
+            {
+                ILogger logger = loggerFactory.CreateLogger<ApplicationInsightsSinkTests>();
+                Dictionary<string, object> telemetryContext = CreateTestTelemetryContext();
+
+                // Act
+                logger.LogEvent(eventName, telemetryContext);
+            }
+
+            // Assert
+            using (ApplicationInsightsDataClient client = CreateApplicationInsightsClient())
+            {
+                await RetryAssertUntilTelemetryShouldBeAvailableAsync(async () =>
+                {
+                    EventsResults<EventsCustomEventResult> results = await client.GetCustomEventsAsync(filter: OnlyLastHourFilter);
+                    Assert.Contains(results.Value, result => result.CustomEvent.Name == eventName);
+                });
+            }
+        }
+
+        [Fact]
+        public async Task LogEventWithComponentName_SinksToApplicationInsights_ResultsInTelemetryWithComponentName()
+        {
+            // Arrange
+            string message = BogusGenerator.Lorem.Sentence();
+            string componentName = BogusGenerator.Commerce.ProductName();
+            using (ILoggerFactory loggerFactory = CreateLoggerFactory(config => config.Enrich.WithComponentName(componentName)))
+            {
+                ILogger logger = loggerFactory.CreateLogger<ApplicationInsightsSinkTests>();
+
+                // Act
+                logger.LogInformation(message);
+            }
+
+            // Assert
+            using (ApplicationInsightsDataClient client = CreateApplicationInsightsClient())
+            {
+                await RetryAssertUntilTelemetryShouldBeAvailableAsync(async () =>
+                {
+                    EventsResults<EventsTraceResult> results = await client.GetTraceEventsAsync(filter: OnlyLastHourFilter);
+                    Assert.NotEmpty(results.Value);
+                    Assert.Contains(results.Value, result => result.Trace.Message == message && result.Cloud.RoleName == componentName);
+                });
+            }
+        }
+
+        [Fact]
+        public async Task LogEventWithVersion_SinksToApplicationInsights_ResultsInTelemetryWithVersion()
+        {
+            // Arrange
+            var eventName = "Update version";
+            using (ILoggerFactory loggerFactory = CreateLoggerFactory(config => config.Enrich.WithVersion()))
+            {
+                ILogger logger = loggerFactory.CreateLogger<ApplicationInsightsSinkTests>();
+
+                // Act
+                logger.LogEvent(eventName);
+            }
+
+            // Assert
+            // Hold on till we have agreed on assertion...
+        }
+
+        [Fact]
+        public void LogEventWithCorrelationInfo_SinksToApplicationInsights_ResultsInTelemetryWithCorrelationInfo()
+        {
+            // Arrange
+            string operationId = $"operation-{Guid.NewGuid()}";
+            string transactionId = $"transaction-{Guid.NewGuid()}";
+            DefaultCorrelationInfoAccessor.Instance.SetCorrelationInfo(new CorrelationInfo(operationId, transactionId));
+            using (ILoggerFactory loggerFactory = CreateLoggerFactory(config => config.Enrich.WithCorrelationInfo()))
+            {
+                ILogger logger = loggerFactory.CreateLogger<ApplicationInsightsSinkTests>();
+
+                // Act
+                logger.LogInformation("Message 1/2 that will be correlated");
+                logger.LogInformation("Message 2/2 that will be correlated");
+
+                // Assert
+                // Hold on till we have agreed on assertion...
+            }
+        }
+
+        [Fact]
+        public void LogEventWithKubernetesInfo_SinksToApplicationInsights_ResultsInTelemetryWithKubernetesInfo()
+        {
+            // Arrange
+            const string kubernetesNodeName = "KUBERNETES_NODE_NAME",
+                         kubernetesPodName = "KUBERNETES_POD_NAME",
+                         kubernetesNamespace = "KUBERNETES_NAMESPACE";
+
+            string nodeName = $"node-{Guid.NewGuid()}";
+            string podName = $"pod-{Guid.NewGuid()}";
+            string @namespace = $"namespace-{Guid.NewGuid()}";
+
+            using (TemporaryEnvironmentVariable.Create(kubernetesNodeName, nodeName))
+            using (TemporaryEnvironmentVariable.Create(kubernetesPodName, podName))
+            using (TemporaryEnvironmentVariable.Create(kubernetesNamespace, @namespace))
+            using (ILoggerFactory loggerFactory = CreateLoggerFactory(config => config.Enrich.WithKubernetesInfo()))
+            {
+                ILogger logger = loggerFactory.CreateLogger<ApplicationInsightsSinkTests>();
+
+                // Act
+                logger.LogInformation("This message will have Kubernetes information");
+                
+                // Assert
+                // Hold on till we have agreed on assertion...
+            }
+        }
+    }
+}
