--- src/Arcus.Observability.Tests.Integration/XunitCollectionBehaviorAttributes.cs
@@ -1,3 +1,0 @@
-﻿using Xunit;
-
-[assembly: CollectionBehavior(CollectionBehavior.CollectionPerAssembly)]
